api/requirements.txt
fastapi
pydantic
pydantic-settings
sqlalchemy
uvicorn
asyncpg
alembic
httpx==0.26.0
pytest
pytest-cov
uvloop
httptools
rich
pyjwt
passlib
sourcery
polars
pytest-asyncio
prometheus-fastapi-instrumentator
alembic
freezegun
prometheus-client